Make API key private
Load API key from a local file that isn't tracked by git. This will make it
private to the public on github. If you want to fork this directory,
currently you must make a file called api_key.txt, put your API key in the
file and nothing else, and put it in the root directory of this project.
Future commits will ask the user for their API key at startup if one is
not provided in api_key.txt.