Change how match history is built
This is a big update. I changed a lot of how the match history gets built.
Details on changes follow.
1. Add method headers that describe the function of each method
succinctly.
2. Update file headers to better reflect what each file does.
3. Change buildMatchHistory file and class to MatchHistoryBuilder
4. In many instances, change how we open files to use python's 'with
open() as f:' functionality, which automatically closes files.
5. Change MatchHistoryBuilder.buildMatch() so that it checks if we have
match details for a certain matchId in match_history_details first, and if
we don't have any details, call getMatchDetails on that matchId. If we get
an error pulling these details, it's likely because we have exceeded our
API key rate limit, so wait 1 second and repeat the method.
6. Take buildMatchHistory out of MatchHistoryBuilder and put it in
MainWindow. This fixes the problem we had early of changing the Gui on a
thread other than the main thread. As a result, we can remove threading
entirely.
7. Change getMatchHistory and getMatchDetails in MatchHistoryBuilder to
just get and return their respective items, and not to write to files.
These methods should function as simply getter methods.
8. MainWindow.buildMatchHistory() now should pass both matchIndex and matchId
to MatchHistoryBuilder.buildMatch().
9. No longer call a worker thread from workerThreads.py. We might use this
file later, but as of now don't use it anymore.

Important note:
If you are first downloading this app, in its current version it won't
stand alone very well. For one thing, you need a file called api_key.txt
with your api key in there for it to run. The app also, on first startup,
holds up the whole UI while it populates match_history_details.txt for the
first time. As of this update, we won't have to populate it every time,
but the first time, it takes upwards of 30-45 seconds to start. After the
first time, it takes around 10 at max. Future updates will address these
issues.