Request API Key from user on login
On login, request the user to input their API Key. This way, someone can
download the app and not know my API Key. If they want to change the API
Key they have entered, they must delete the config.ini file in the
directory where the app is located and run the app again. The app will
also check in the directory for a file called api_key.txt, which should
contain the key, if there is a problem loading the key from the config file.