Make initial app startup smoother
When starting the app for the first time, we pull detailed match data on
the user's entire match history for the 2016 season. During this time,
make the app responsive to the user and display a progress dialog that
shows them how far along we are in the downloading process. This behavior
should only happen if we don't have any match history data yet (that is,
match_history.txt doesn't exist).