Produce an executable file for the application
Build an executable file called main.exe using pyinstaller. This creates
the /dist directory, in which it builds an executable version of this
project. If you just want to run the program, clone the repo and enter the
/dist/main directory, then run main.exe. Creating an executable greatly
simplifies the process a user goes through if they simply want to use the
program, but not use the code.

On my local machine (Windows), whitespace is configured naturally as CRLF.
This means that if you want to run the program on a linux machine, it
might not work. I'll decide soon whether to change this, but as of now,
the application works on Windows.