Resolve ui file error in app executable
The app expected the .ui file to be in a hardcoded location. The app now
looks for it in the working directory, no matter its path.
Note: MainWindow.ui does not get automatically added to the dist/main
directory with the 'pyinsaller main.py' command. One must move it into
this directory before running the app.