Add Summoner class to hold summoner information
Details:
1. Modify the way we hold summoner information from globals to variables
tied to an instance of the Summoner class. This is cleaner, easier to
track, and opens the possibility of multiple summoners for things like
summoner comparisons.
2. Add updateMatchHistoryVariables method and use it to send data between
classes/threads. This method should be called once match history worker
threads finish. Create worker thread instances with the match history
variables as arguments.
3. Fix refresh button to correctly refresh match history, summoner
information, and the gui.
4. Clean up closeEvent by ensuring the app closes if the user doesn't
enter a summoner name or api key on first login and ensuring the app
correctly writes match history variables to files.
5. Remove extraneous checks for api_key.txt, as they aren't used.