Calculate basic match statistics and show them
Implement the calculateMatchStatistics and calculateAverages methods in
the MatchHistoryBuilder class so that we can see the user's score, kda,
kill participation percentage, ward score, gold per minute, and cs per
minute for each match. Make these values red or green depending on if they
are lower or higher than the user's average, respectively.