Fix bug in first match of match history
When first downloading champion name-id pairs, the app was using the last
downloaded champion ID instead of the ID passed to the function. Go
figure.

Also, this commit should update the dist/ and /build directories with a
new executable. This was forgotten last commit. I found a way to include
files like MainWindow.ui and about.txt that didn't get automatically
compiled in the /dist folder, so those should be a part of future commits
reliably.