Implement calculateAverages method
1. Implement the calculateAverages method in the MatchHistoryBuilder class
so that it loops through all matchHistoryStatistics for each match and
calculates an average for the last 20 games and another for all games.
2. Change the calculateMatchStatistics method to be calculateStatistics,
and instead of calling this method for every match, call it once in
buildMatchHistory and loop through all matches, calculating their
statistics, and storing them in matchHistoryStatistics.
3. I also included some preliminary code testing out how to gracefully
kill threads, but haven't gotten it to work yet. These changes can be
disregarded for now, but this commit should show that they happen here.